Create A - WATERMELON
Given the the size of the watermelon that will be an even number.
After dividing, the size must be even. Hence 2 is the only even number that becomes odd when divided evenly. Give the condition as it must not be 2 and it must be even and print yes or no accordingly.